Using pytthon to create a html
I used to write a python to finish a tesing . But just a single text report makes it not easy to see the result. And a HTML Website can use different color , different word size and show the detail functions to make report more easy to show the result and check the mistake.
This is the example for the HTML report to show how it works.